requirements.txt
streamlit
python-docx
pytesseract
easyocr
opencv-python
PyPDF2
pdfplumber
moviepy
ffmpeg-python
openai-whisper
SpeechRecognition
pydub
transformers
torch
sentencepiece
